First

+++ glyke_back/views.py
@@ -33,7 +33,7 @@ def get_category_fields(*, category=None):
     category_fields = {}
     if category:
         for filter in category.filters.all(): # these are django-taggit objects
-                    # TODO Add some king of validation here to select the right input field
+                    # TODO Add some kind of validation here to select the right input field
                     category_fields[str(filter)] = forms.CharField(widget=forms.TextInput(attrs={'class': 'form-control',},),
                                                                                           label=_(str(filter).capitalize()))
     return category_fields
@@ -51,6 +51,14 @@ def get_photo_from_img_src(*, img_src, model_instance, exclude = '_display'):
     photo_image_substr = get_photo_image_name_from_img_src(img_src, exclude=exclude)
     photo_queryset = model_instance.photos.photos.filter(image__endswith = photo_image_substr)
     if photo_queryset.count() == 1: return photo_queryset.first()
+
+def get_order(request, *, status, order_by='-created'):
+    """Checks if there is an order of passed status, redirects to error 500 if not
+    status: status str as stated in model definition
+    order_by: generic django's ordering variable [default='-created']"""
+    if not request.user.orders.filter(status=status).exists(): # check if there is a 'current' order
+        return redirect(f"{reverse('smth_went_wrong')}?{urlencode({'error_suffix': 'order (probably there is none)'})}")
+    return request.user.orders.filter(status=status).order_by(order_by).first()
 
 @user_is_staff_or_404()
 @require_http_methods(["GET", "POST"])
@@ -269,13 +277,13 @@ class SignInView(LoginView):
     authentication_form = SignInForm
     template_name = 'sign_in.html'
 
+
 @require_http_methods(["GET", "POST"])
 def cart_view(request):
-
+    current_order = get_order(request, status='CUR') # select the latest current order
     if request.method=='POST':
         products_id_set = set(request.POST.getlist('products_id'))
-        current_order = request.user.orders.all().first()
-        current_order.order_lines.all().delete()
+        if products_id_set: current_order.order_lines.all().delete()
 
         for product_id in products_id_set:
             quantity_list = request.POST.getlist(f'quantity_{product_id}')
@@ -289,22 +297,12 @@ def cart_view(request):
                                      quantity = new_quantity,
                                      )
 
-        # current_order.order_lines.filter(product=Product.objects.get(id=request.POST['product_id']))
-
-
-
 
     context = {}
 
-
-
-
-    context['order'] = request.user.orders.all().first()
-    if request.user.orders.all().exists():
-        context['order_lines'] = request.user.orders.all().first().order_lines.all()
+    context['order'] = current_order
+    context['order_lines'] = current_order.order_lines.all()
     return render(request, "cart.html", context)
-
-
 
 
 class AddToCartView(LoginRequiredMixin, RedirectView):
@@ -315,22 +313,15 @@ class AddToCartView(LoginRequiredMixin, RedirectView):
     def dispatch(self, request, *args, **kwargs):
         # from LoginRequiredMixin.dispatch
         if not request.user.is_authenticated: return self.handle_no_permission()
-
-        if not request.user.orders.filter(status='CUR').exists(): # check if there is a 'current' order
-            self.url = f"{reverse('smth_went_wrong')}?{urlencode({'error_suffix': 'current order'})}"
-        else:
-            self.url = request.POST.get('next', '/') # set redirect url
-            current_order = request.user.orders.filter(status='CUR').order_by('-created').first()
-            product_id = request.POST.get('product_id')
-            OrderLine.objects.create(parent_order=current_order,
-                                     product=Product.objects.get(id=product_id),
-                                     )
+        # from RedirectView.dispatch
+        if not request.method.lower() in self.http_method_names: return self.http_method_not_allowed(request, *args, **kwargs)
+        current_order = get_order(request, status='CUR') # select the latest current order
+        self.url = request.POST.get('next', '/') # set redirect url
+        product_id = request.POST.get('product_id')
+        OrderLine.objects.create(parent_order=current_order,
+                                 product=Product.objects.get(id=product_id),
+                                 )
         return RedirectView.dispatch(self, request, *args, **kwargs)
 
 
-
-
-
-
-# TODO add cart view
-# TODO add cart header (add to user header??)
+# TODO something has to handler current_order creation before add_to_cart's and cart's views
